Remove redundant level check
