+++ src/Bot.cs
@@ -36,6 +36,7 @@
     // TODO: Subscriptions via DM again
     // TODO: Multiple discord bot tokens per server
     // TODO: Custom gender emojis
+    // TODO: Double check gym alarms
 
     public class Bot
     {
